requirements.txt
# Core Application Dependencies
streamlit==1.33.0
openai==1.16.2
dspy-ai==2.4.3
python-dotenv
sqlalchemy
jsonschema
python-json-logger
psycopg2-binary
bleach

# Data & Serialization
cloudpickle>=3.0.0
pandas>=2.1.4
numpy>=1.26.3

# Utilities
cachetools>=5.3.2
typing-extensions>=4.9.0
aiohttp>=3.9.1
